Fixes in makefile

+++ mlops/detection_orchestration/charts/feature_profiles_for_ingest.py
@@ -8,7 +8,16 @@ df_1 = convert_matrix_to_dataframe(df_1)
 df_1 = df_1.iloc[:, :DATAFRAME_ANALYSIS_MAX_COLUMNS]
 columns_and_types = infer_column_types(df_1).items()
 columns = [t[0] for t in columns_and_types]
-stats = ['Type', 'Missing values', 'Unique values', 'Min', 'Max', 'Mean', 'Median', 'Mode']
+stats = [
+    "Type",
+    "Missing values",
+    "Unique values",
+    "Min",
+    "Max",
+    "Mean",
+    "Median",
+    "Mode",
+]
 rows = [[] for _ in stats]
 
 for col, col_type in columns_and_types:
@@ -24,14 +33,14 @@ for col, col_type in columns_and_types:
     if len(not_null) == 0:
         continue
 
-    if col_type.value in ['number', 'number_with_decimals']:
-        if str(series.dtype) == 'object':
-            if col_type.value == 'number_with_decimals':
-                series = series.astype('float64')
-                not_null = not_null.astype('float64')
+    if col_type.value in ["number", "number_with_decimals"]:
+        if str(series.dtype) == "object":
+            if col_type.value == "number_with_decimals":
+                series = series.astype("float64")
+                not_null = not_null.astype("float64")
             else:
-                series = series.astype('int64')
-                not_null = not_null.astype('int64')
+                series = series.astype("int64")
+                not_null = not_null.astype("int64")
 
         count = len(not_null.index)
         if count >= 1:
@@ -44,18 +53,20 @@ for col, col_type in columns_and_types:
         max_value = not_null.astype(str).max()
 
     _, mode = sorted(
-      [(v, k) for k, v in not_null.value_counts().items()],
-      reverse=True,
+        [(v, k) for k, v in not_null.value_counts().items()],
+        reverse=True,
     )[0]
 
-    for idx, value in enumerate([
-        col_type.value,
-        len(series[series.isna()].index),
-        len(series.unique()),
-        min_value,
-        max_value,
-        mean,
-        median,
-        mode,
-    ]):
-      rows[idx].append(value)
+    for idx, value in enumerate(
+        [
+            col_type.value,
+            len(series[series.isna()].index),
+            len(series.unique()),
+            min_value,
+            max_value,
+            mean,
+            median,
+            mode,
+        ]
+    ):
+        rows[idx].append(value)
